0914送るやつ

+++ uriagekanri/src/main/java/com/example/demo/entity/Clientname.java
@@ -13,13 +13,13 @@ import javax.persistence.Table;
 public class Clientname implements Serializable{
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Long nameid;
+	private int nameid;
 
-	public Long getNameid() {
+	public int getNameid() {
 		return nameid;
 	}
 
-	public void setNameid(Long nameid) {
+	public void setNameid(int nameid) {
 		this.nameid = nameid;
 	}
