+++ uriagekanri/src/main/java/com/example/demo/repository/Client2Repository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Client2Ste;
 
 @Repository
-public interface Client2Repository extends JpaRepository<Client2Ste, Long>{
+public interface Client2Repository extends JpaRepository<Client2Ste, Integer>{
 
 	@Query(value="SELECT * FROM client2",nativeQuery=true)
 	List<Client2Ste> clientSte2();
