+++ uriagekanri/src/main/java/com/example/demo/repository/Client3Repository.java
@@ -10,7 +10,7 @@ import com.example.demo.entity.Client3Ste;
 
 
 @Repository
-public interface Client3Repository extends JpaRepository<Client3Ste, Long>{
+public interface Client3Repository extends JpaRepository<Client3Ste, Integer>{
 
 	@Query(value="SELECT * FROM client3",nativeQuery=true)
 	List<Client3Ste> clientSte3();
