+++ uriagekanri/src/main/java/com/example/demo/repository/ListRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.entity.Listdata;
 
 @Repository
-public interface ListRepository extends JpaRepository<Listdata, Long>{
+public interface ListRepository extends JpaRepository<Listdata, Integer>{
 	@Query(value="SELECT id,maindata.nameid,orderdate,snumber,subtitle,items,delivery_designated_date,delivery_date,billing_date,estimated_amount,order_amount,statusid,remark,delete_flg,clientname.nameid,client FROM maindata INNER JOIN clientname ON maindata.nameid=clientname.nameid WHERE delete_flg=0 order by maindata.id ASC",nativeQuery=true)
 	Page<Listdata> findAlldataA(Pageable pageable);
